Refactored code and file names

+++ app/src/cpp/line-sensor/LineSensor.cpp
@@ -1,12 +1,12 @@
 #include "LineSensor.h"
 
-LineSensor::LineSensor(int sensorPin)
+LineSensor::LineSensor(uint8_t sensorPin)
     : pin(sensorPin)
 {
     pinMode(sensorPin, INPUT);
 }
 
-void LineSensor::setReading(int reading)
+void LineSensor::setReading(uint8_t reading)
 {
     if (reading == 0 || reading == 1)
         internalReading = reading;
